move 'aperture' tests to separate file
